Split the node network layer out of the node kernel
This should be a pure refactoring, with no changes. It takes the network
interface initialisation out of the NodeKernel and moves it into its own
layer which can be constructed from the NodeKernel.

This is another step in the direction of fully integrating the remaining
network layer features (mux, version negotiation etc).
