+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -83,6 +83,7 @@ library
                        Ouroboros.Consensus.Mempool.Impl
                        Ouroboros.Consensus.Node
                        Ouroboros.Consensus.NodeId
+                       Ouroboros.Consensus.NodeNetwork
                        Ouroboros.Consensus.Protocol.Abstract
                        Ouroboros.Consensus.Protocol.BFT
                        Ouroboros.Consensus.Protocol.ExtNodeConfig
